requirements.txt
# Python dependencies for CryptoBot
Flask==2.2.3
Flask-Cors==3.0.10
python-binance==1.0.16
pandas==1.5.3
numpy==1.24.2
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
requests==2.28.2
python-dotenv==1.0.0
PyJWT==2.6.0
colorama==0.4.6
pytz==2023.3
websocket-client==1.5.1
cryptography==40.0.2
